LayoutTests/platform/chromium-win/ietestcenter/Javascript/15.5.4.20-1-1-expected.txt

15.5.4.20-1-1 (TestCases/chapter15/15.5/15.5.4/15.5.4.20/15.5.4.20-1-1.js) String.prototype.trim throws TypeError when string is undefined

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ES5Harness.preconditionPassed is true
FAIL ES5Harness.testPassed should be true (of type boolean). Was undefined (of type undefined).
PASS successfullyParsed is true

TEST COMPLETE
